# src/main/resources/openapi/notification-api.yaml
openapi: 3.0.3
info:
  title: Group 8 Communication and Feedback API
  version: 1.0.0
  description: |
    Portable contract for in-app notifications, announcements, feedback and engagement.
    The running service also generates documentation from controllers at /v3/api-docs
    and /v3/api-docs.yaml, with interactive documentation at /swagger-ui.html.
    User operations require a Group 5 RS256 token; trusted triggers use X-Service-Key.
    Notification creation stores an inbox record, not an email or SMS.
servers:
  - url: /
paths:
...
  /api/announcements:
    get:
      summary: List published announcements visible to the authenticated user
      description: Falls back to ALL and JWT ROLE audiences if Group 5 is unconfigured or unavailable.
      operationId: listVisibleAnnouncements
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Published announcements matching the authenticated user's Group 5 audience data.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Announcement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Authenticated user was not found in Group 5 directory.
    post:
      summary: Create an announcement draft
      description: Requires ADMIN or ADMINISTRATIVE_STAFF. ALL omits ruleValue; other audiences require it.
      operationId: createAnnouncement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreateRequest'
      responses:
        '201':
          description: Draft created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '400':
          description: Invalid announcement or targeting rule.
        '403':
          description: Requires ADMIN or ADMINISTRATIVE_STAFF.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/announcements/{id}/publish:
    post:
      summary: Publish an announcement draft
      description: Requires ADMIN or ADMINISTRATIVE_STAFF and ownership of the draft.
      operationId: publishAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AnnouncementId'
      responses:
        '200':
          description: Announcement published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The authenticated user did not create the announcement.
        '404':
          description: Announcement was not found.
        '409':
          description: Announcement is not a draft.
  /api/announcements/{id}/archive:
    post:
      summary: Archive a published announcement
      description: Requires ADMIN or ADMINISTRATIVE_STAFF and ownership of the announcement.
      operationId: archiveAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AnnouncementId'
      responses:
        '200':
          description: Announcement archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The authenticated user did not create the announcement.
        '404':
          description: Announcement was not found.
        '409':
          description: Announcement is not published.
  /api/feedback/forms:
    get:
      summary: List active feedback forms
      operationId: listFeedbackForms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active forms ordered newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackForm'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a feedback form for an event or service request
      description: Requires EVENT_ORGANIZER, ACADEMIC_STAFF, ADMINISTRATIVE_STAFF or ADMIN.
      operationId: createFeedbackForm
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackFormCreateRequest'
      responses:
        '201':
          description: Feedback form created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackForm'
        '400':
          description: Invalid feedback form.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The authenticated user does not have a form-creation role.
        '409':
          description: A form already exists for the activity.
  /api/feedback/forms/{formId}:
    get:
      summary: Get one feedback form
      operationId: getFeedbackForm
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeedbackFormId'
      responses:
        '200':
          description: Feedback form.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackForm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Feedback form was not found.
  /api/feedback/forms/{formId}/responses:
    get:
      summary: List responses for a feedback form
      description: Only the form creator can read responses.
      operationId: listFeedbackResponses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeedbackFormId'
      responses:
        '200':
          description: Feedback responses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The authenticated user did not create the form.
        '404':
          description: Feedback form was not found.
    post:
      summary: Submit one response to a feedback form
      description: |
        Requires an active form and an eligible caller. Group 7 service requests must be
        RESOLVED or CLOSED and owned by the caller. Event feedback needs a CONFIRMED
        registration. A blank provider base URL skips its check for development only.
      operationId: submitFeedbackResponse
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeedbackFormId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackResponseRequest'
      responses:
        '201':
          description: Feedback response created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          description: Rating or response data is invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The user is not eligible to submit feedback.
        '404':
          description: Feedback form was not found.
        '409':
          description: Form is inactive or the user already submitted a response.
        '503':
          description: Group 7 or event-service eligibility provider is unavailable or returned invalid data.
  /api/engagement-dashboard/summary:
    get:
      summary: Get communication and feedback engagement summary
      operationId: getEngagementSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary metrics owned by the communication-feedback service.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: X-Service-Key
  parameters:
...
    AnnouncementId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FeedbackFormId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
  schemas:
...
    AudienceType:
      type: string
      enum:
        - ALL
        - ROLE
        - DEPARTMENT
        - FACULTY
        - SERVICE_UNIT
    AnnouncementCreateRequest:
      type: object
      required:
        - title
        - content
        - audienceType
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
        audienceType:
          $ref: '#/components/schemas/AudienceType'
        ruleValue:
          type: string
          maxLength: 100
    Announcement:
      type: object
      required:
        - id
        - title
        - content
        - status
        - createdBy
        - audienceType
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        createdBy:
          type: string
          maxLength: 64
        audienceType:
          $ref: '#/components/schemas/AudienceType'
        ruleValue:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FeedbackFormCreateRequest:
      type: object
      required:
        - activityType
        - activityId
        - title
        - questionsJson
      properties:
        activityType:
          type: string
          enum: [EVENT, SERVICE_REQUEST]
        activityId:
          type: string
          maxLength: 64
          description: Event id (UUID) or Group 7 service request id, e.g. REQ-2026-004
        title:
          type: string
          maxLength: 200
        questionsJson:
          type: string
          minLength: 1
          maxLength: 20000
          description: JSON question definition encoded as a string, not an array property.
    FeedbackForm:
      type: object
      required:
        - id
        - activityType
        - activityId
        - createdBy
        - title
        - questionsJson
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        activityType:
          type: string
          enum: [EVENT, SERVICE_REQUEST]
        activityId:
          type: string
          maxLength: 64
          description: Event id (UUID) or Group 7 service request id, e.g. REQ-2026-004
        createdBy:
          type: string
          maxLength: 64
        title:
          type: string
        questionsJson:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FeedbackResponseRequest:
      type: object
      required:
        - rating
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 2000
    FeedbackResponse:
      type: object
      required:
        - id
        - formId
        - activityId
        - respondentId
        - rating
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        formId:
          type: string
          format: uuid
        activityId:
          type: string
          maxLength: 64
          description: Event id (UUID) or Group 7 service request id, e.g. REQ-2026-004
        respondentId:
          type: string
          maxLength: 64
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    EngagementSummary:
      type: object
      required:
        - publishedAnnouncementCount
        - activeFeedbackFormCount
        - feedbackResponseCount
        - notificationCount
        - eventParticipationAvailable
      properties:
        publishedAnnouncementCount:
          type: integer
          format: int64
        activeFeedbackFormCount:
          type: integer
          format: int64
        feedbackResponseCount:
          type: integer
          format: int64
        averageFeedbackRating:
          type: number
          format: double
          nullable: true
        notificationCount:
          type: integer
          format: int64
        eventParticipationCount:
          type: integer
          format: int64
          nullable: true
        eventParticipationAvailable:
          type: boolean